Don't call into GTK+ when setting video sink from gstreamer thread
As noted at http://bugs.sugarlabs.org/ticket/2853
Jukebox often crashes when loading content or shows its content in the
wrong window.

This is because it is calling into GTK+ from the gstreamer thread,
without taking the appropriate locks.

Fix this by following the suggestion of the gstreamer documentation, where
the xid is queried earlier from the correct thread:
http://gstreamer.freedesktop.org/data/doc/gstreamer/head/gst-plugins-base-libs/html/gst-plugins-base-libs-gstxoverlay.html