Update time and state on seek
When the horizontal scroll changes its value it updates the time
position (label) to know where are we going to drop the stream.

The state of the stream[1] is updated with a timeout of 1ns, so it
doesn't consume too much CPU. This impacts directily the performance
on seeking videos and audios.

[1] http://gstreamer.freedesktop.org/data/doc/gstreamer/head/gstreamer/html/GstElement.html#gst-element-get-state

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>