Create GstPlayer just once
Instanciating GstPlayer when Jukebox is launched and changing its uri
when a new stream is played. This allow us to change between different
streams in the playlist without having many memory issues regarding
video playing in HD.

To change the stream being played in the GstPlayer instance we firstly
set the state of it to STATE_PAUSED, then to STATE_NULL and finally we
change the uri. When this process finishes we change the state again
to STATE_PLAYING.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>