Use 'playbin2' instead of 'playbin'
'playbin' is obsolete and not should be used. I was told in #gstreamer
(irc.freenode.net) that 'playbin' is no longer supported, that has
some bugs and shouldn't be used anymore. It should be changed by
'playbin2'

I had to set some flags as well to keep showing the 'vis-plugin' with
'playbin2' player.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>