Check availability of stream
It's possible to have some streams in the playlist that are not
available at the moment the user double-clicks the row or presses the
Play button, for example because they are in a Pen Drive.

So, this patch checks for the availability of the stream before start
playing it and avoiding gstreamer to fail.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>