Fix: 'Add stream from the Journal (datastore)'
We need to save the 'object_id' for those streams that are inside the
datastore because the path is removed every time that Sugar is
restarted. So, we need the 'object_id' value to get the real path
every time that we are about to play the stream.

This bug was introduced in 8f2924e8857cdce725b8c2b2cd64dfb618a28210

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>