Syncronize element been playerd with selection in the treeview
In PlayList class code, two properties were used to track
the item been playes, _current_playing and the method set_cursor
to set the position in the treeview. As _current_playing was modified
outside of the class, they lost sync. Methods set_curren_playing() and
get_current_playing are added and the private prop is not used anymore
outside of the class. The method set_cursor is now private.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>